Refactored logstash converter mapping access modifiers
